allow custom terrain generator
