+++ source/isaaclab/isaaclab/terrains/height_field/hf_terrains_cfg.py
@@ -7,7 +7,7 @@ from dataclasses import MISSING
 
 from isaaclab.utils import configclass
 
-from ..terrain_generator_cfg import SubTerrainBaseCfg
+from ..sub_terrain_cfg import SubTerrainBaseCfg
 from . import hf_terrains
