+++ source/isaaclab/isaaclab/terrains/terrain_importer.py
@@ -16,7 +16,6 @@ import isaaclab.sim as sim_utils
 from isaaclab.markers import VisualizationMarkers
 from isaaclab.markers.config import FRAME_MARKER_CFG
 
-from .terrain_generator import TerrainGenerator
 from .utils import create_prim_from_mesh
 
 if TYPE_CHECKING:
@@ -84,7 +83,7 @@ class TerrainImporter:
             if self.cfg.terrain_generator is None:
                 raise ValueError("Input terrain type is 'generator' but no value provided for 'terrain_generator'.")
             # generate the terrain
-            terrain_generator = TerrainGenerator(cfg=self.cfg.terrain_generator, device=self.device)
+            terrain_generator = self.cfg.terrain_generator.class_type(cfg=self.cfg.terrain_generator, device=self.device)
             self.import_mesh("terrain", terrain_generator.terrain_mesh)
             # configure the terrain origins based on the terrain generator
             self.configure_env_origins(terrain_generator.terrain_origins)
