+++ source/isaaclab/isaaclab/terrains/trimesh/mesh_terrains_cfg.py
@@ -10,7 +10,7 @@ import isaaclab.terrains.trimesh.mesh_terrains as mesh_terrains
 import isaaclab.terrains.trimesh.utils as mesh_utils_terrains
 from isaaclab.utils import configclass
 
-from ..terrain_generator_cfg import SubTerrainBaseCfg
+from ..sub_terrain_cfg import SubTerrainBaseCfg
 
 """
 Different trimesh terrain configurations.
